+++ sentry/src/main/java/io/sentry/NoOpTransaction.java
@@ -76,13 +76,13 @@ public final class NoOpTransaction implements ITransaction {
   }
 
   @Override
-  public @NotNull TraceState traceState() {
-    return new TraceState(SentryId.EMPTY_ID, "");
+  public @NotNull TraceContext traceContext() {
+    return new TraceContext(SentryId.EMPTY_ID, "");
   }
 
   @Override
-  public @NotNull TraceStateHeader toTraceStateHeader() {
-    return new TraceStateHeader("");
+  public @NotNull BaggageHeader toBaggageHeader() {
+    return new BaggageHeader("");
   }
 
   @Override